test organization and cleanup
